Capture pixel colour at tapped target

+++ app/src/main/java/com/danielhood/rotationcounter/CounterViewModel.kt
@@ -7,8 +7,8 @@ import android.view.View
 class CounterViewModel () {
     var rotationCount: Int = 0
     var currentFps: Float = 0f
-    var targetX: Float = 0f
-    var targetY: Float = 0f
+    var targetX: Int = 0
+    var targetY: Int = 0
 
     val rotationCountString: String
         get()  = rotationCount.toString()
